requirements.txt
requests
apprise
dnspython